+++ cli/fledger/src/main.rs
@@ -8,7 +8,7 @@ use flarch::{
 use flmodules::{
     dht_router::broker::DHTRouter,
     dht_storage::broker::DHTStorage,
-    network::{broker::NetworkIn, network_start, signal::SIGNAL_VERSION},
+    network::{broker::NetworkIn, signal::SIGNAL_VERSION},
 };
 use flnode::{node::Node, version::VERSION_STRING};
 use page::{Page, PageCommands};
@@ -134,8 +134,7 @@ async fn main() -> anyhow::Result<()> {
         )
     };
     log::debug!("Connecting to websocket at {:?}", cc);
-    let network = network_start(node_config.clone(), cc).await?;
-    let node = Node::start(Box::new(storage), node_config, network.broker).await?;
+    let node = Node::start_network(Box::new(storage), node_config, cc).await?;
     Fledger::run(node, args).await
 }